add foundation for rocketfuel mock API
